requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-dateutil
psycopg2-binary